Remove git-annex from bash_exports
If git-annex is sourced from a package that is not part of the package
manager of the distribution, it will bring its own Git installation.
This might conflict with other tools like pre-commit as some paths seem
to be different from what is expected. Thus, git-annex should rather be
installed via a submodule (modulefiles/lmod) instead of always sourcing
it.