Add fix make GTK+ app start faster on swaywm
This is based on the explanation in the sway wiki [1]:

```text
GTK+ applications take 20 seconds to start

This is due to GTK+ waiting for xdg-desktop-portal to start via D-Bus.
This times out because the D-Bus activated service doesn't know what
WAYLAND_DISPLAY to connect to.

This can be fixed by adding the following to your configuration file:

exec dbus-update-activation-environment --systemd DISPLAY
WAYLAND_DISPLAY SWAYSOCK

Alternatively, set GTK_USE_PORTAL=0 in your environment.
```

[1]:
https://github.com/swaywm/sway/wiki#gtk-applications-take-20-seconds-to-start